refactor: move filter components to submissions page
