+++ packages/frontend/src/app/submissions/page.tsx
@@ -1,14 +1,24 @@
 'use client'
-import React, { useEffect, useState } from 'react'
+import React, { ReactNode, useEffect, useState } from 'react'
 import Image from 'next/image'
 import dayjs from 'dayjs'
 import { PageHeading } from '@/common/components/PageHeading'
-import { Text4xlBold } from '@/common/components/Typography'
+import { Text4xlBold, TextSmBold } from '@/common/components/Typography'
 import { Submission } from '@/common/interfaces'
 import { PageContainer } from '@/common/components/PageContainer'
 import { Table, TableBody, TableHeader } from '@/common/components/Table'
-import { Filter, FilterButton } from '@/common/components/Filter'
 import { betterFetch } from '@/api'
+import { PlusIcon } from '@/common/components/images/PlusIcon'
+
+interface FilterProps {
+  children: ReactNode
+}
+
+interface FilterButtonProps {
+  onClick: any
+  isActive: boolean
+  value: string
+}
 
 export default function Page() {
   const [submissions, setSubmissions] = useState([])
@@ -127,5 +137,36 @@ export default function Page() {
         </TableBody>
       </Table>
     </PageContainer>
+  )
+}
+
+function Filter({ children }: FilterProps) {
+  return (
+    <div className="flex flex-col">
+      <TextSmBold className="ml-4 h-6 mb-2">Filters</TextSmBold>
+      <div className="flex gap-x-2 mb-6">{children}</div>
+    </div>
+  )
+}
+
+function FilterButton({ onClick, isActive, value }: FilterButtonProps) {
+  return (
+    <button
+      onClick={onClick}
+      className={
+        'btn rounded-full min-h-0 h-10 text-sm normal-case transition ' +
+        (isActive
+          ? 'bg-primary hover:bg-primary text-white'
+          : 'bg-white hover:bg-white text-primary')
+      }
+    >
+      <PlusIcon
+        className={
+          'transition duration-500 ' +
+          (isActive ? 'rotate-45 text-white' : 'text-primary')
+        }
+      />
+      <p className="capitalize">{value}</p>
+    </button>
   )
 }
